Revert "use ES6 import and export"
This reverts commit 04f4607922e84d93ca8a05708802fb6bbd61f359.
