+++ ext/bg/js/search.js
@@ -16,7 +16,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {Display} from '../../mixed/js/display.js';
 
 class DisplaySearch extends Display {
     constructor() {
