+++ ext/bg/search.html
@@ -43,8 +43,9 @@
         <script src="/bg/js/handlebars.js"></script>
         <script src="/bg/js/templates.js"></script>
         <script src="/bg/js/util.js"></script>
+        <script src="/mixed/js/display.js"></script>
         <script src="/mixed/js/japanese.js"></script>
 
-        <script type="module" src="/bg/js/search.js"></script>
+        <script src="/bg/js/search.js"></script>
     </body>
 </html>
