+++ ext/fg/float.html
@@ -36,7 +36,10 @@
 
         <script src="/fg/js/api.js"></script>
         <script src="/fg/js/util.js"></script>
+        <script src="/fg/js/document.js"></script>
+        <script src="/fg/js/source.js"></script>
+        <script src="/mixed/js/display.js"></script>
 
-        <script type="module" src="/fg/js/float.js"></script>
+        <script src="/fg/js/float.js"></script>
     </body>
 </html>
