+++ ext/fg/js/float.js
@@ -16,7 +16,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {Display} from '../../mixed/js/display.js';
 
 class DisplayFloat extends Display {
     constructor() {
@@ -122,6 +121,41 @@ class DisplayFloat extends Display {
 
         if (this.styleNode.parentNode !== parent) {
             parent.appendChild(this.styleNode);
+        }
+    }
+
+    async onTermLookup(e) {
+        try {
+            e.preventDefault();
+
+            const clickedElement = $(e.target);
+            const textSource = docRangeFromPoint({x: e.clientX, y: e.clientY});
+            textSource.setEndOffset(this.options.scanning.length);
+
+            const {definitions, length} = await apiTermsFind(textSource.text());
+            if (definitions.length === 0) {
+                return false;
+            }
+
+            textSource.setEndOffset(length);
+
+            const sentence = docSentenceExtract(textSource, this.options.anki.sentenceExt);
+
+            const context = {
+                source: {
+                    definitions: this.definitions,
+                    index: Display.entryIndexFind(clickedElement)
+                }
+            };
+
+            if (this.context) {
+                context.sentence = sentence;
+                context.url = this.context.url;
+            }
+
+            this.termsShow(definitions, this.options, context);
+        } catch (e) {
+            this.onError(e);
         }
     }
 }
