+++ ext/mixed/js/display.js
@@ -17,8 +17,6 @@
  */
 
 
-import {docRangeFromPoint, docSentenceExtract} from '../../fg/js/document.module.js';
-
 class Display {
     constructor(spinner, container) {
         this.spinner = spinner;
@@ -66,41 +64,6 @@ class Display {
 
             const kanjiDefs = await apiKanjiFind(link.text());
             this.kanjiShow(kanjiDefs, this.options, context);
-        } catch (e) {
-            this.onError(e);
-        }
-    }
-
-    async onTermLookup(e) {
-        try {
-            e.preventDefault();
-
-            const clickedElement = $(e.target);
-            const textSource = docRangeFromPoint({x: e.clientX, y: e.clientY});
-            textSource.setEndOffset(this.options.scanning.length);
-
-            const {definitions, length} = await apiTermsFind(textSource.text());
-            if (definitions.length === 0) {
-                return false;
-            }
-
-            textSource.setEndOffset(length);
-
-            const sentence = docSentenceExtract(textSource, this.options.anki.sentenceExt);
-
-            const context = {
-                source: {
-                    definitions: this.definitions,
-                    index: Display.entryIndexFind(clickedElement)
-                }
-            };
-
-            if (this.context) {
-                context.sentence = sentence;
-                context.url = this.context.url;
-            }
-
-            this.termsShow(definitions, this.options, context);
         } catch (e) {
             this.onError(e);
         }
@@ -497,5 +460,3 @@ class Display {
         return $('.entry').eq(index).find('.action-view-note');
     }
 }
-
-export {Display};
